refactor: move some components around, clean up InnerModal
