+++ frontend/src/components/modal/utils/useIsRunningOnClient.ts
